requirements.txt
typer[all]
supabase~=2.15.3
python-dotenv~=1.1.0